Put `stripe_test.go` into main `stripe` package
Moves `stripe_test.go` out of the `stripe_test` package and into
`stripe`. This makes it consistent with all other `*_test.go` files in
the top-level package. The up side of the change is that it will allow
us to test certain unexported functions without having to export them.
The down side is that the hygiene isn't quite as good because we're no
longer just testing against the package's public interface which may
reduce flexibility in future changes.

Go standard library convention seems to use both methods. Just looking
through the `net/http` source code for example, you can find tests in
the `http` package _and_ in the `http_test` package depending on the
file.

I've been wanting to making this change for a while because it'd give us
a route to making internal-ish structures like `BackendImplementation`
unexported, but the direct impetus for today is that I was making some
changes to the `shouldRetry` function and noticed that it was difficult
to test them without having to go through the expensive song and dance
of spinning up a test HTTP server and making the test really
complicated. (I could also just export the function as `ShouldRetry`,
but I'd prefer not to do that if possible.)