Rearrange error logging so that 402 doesn't log an error
While looking in #936 I accidentally realized that in #935 I'd
inadvertently duplicated error logging -- I add a log statement in `Do`
for API errors, but `ResponseToError` was already logging them.

Here I've resolved #936 and re-arranged logging somewhat so that we're
more consistent _where_ the logging happens. Most logs are now produced
directly in `Do` so it's easier to follow what's getting logged.

I haven't actually changed our previous policy around 402s -- just moved
it up a function. The previous idea (which I'd forgotten about) was that
because 402s are often induced by the _users of integrations_ instead of
integrations themselves, we just do an `Infof` log for it. Other errors
get an `Errorf` because they're more likely to be the integration's
fault, and actionable.

Fixes #936.