Also log error in case of non-`stripe.Error`
In a few cases it's possible for `ResponseToError` to return an error
that's not a `*stripe.Error`. This should never happen, but could occur
if say the Stripe API returned a mangled JSON blob, but somehow managed
to close the connection properly. It's an unlikely scenario, but even
so, it's better not to let it slip through the cracks.

I also moved one more logging statement out of `ResponseToError` so we
can log everything of this type in one place.