Remove lossy "MarshalJSON" implementations
See #946 for context, but we're removing some `MarshalJSON`
implementations that are "lossy" in that they throw out information that
might be useful to somebody.

Note that we don't actually marshal API resources as part of normal
operations in stripe-go, so the only users that this will affect are
those that are manually re-marshaling resources to JSON in their
integrations.