Fix top-level `DeclineCode` deserialization in errors
Issue #1115 reveals a bug with `DeclineCode` on the top-level error
object.

In #949, we added `DeclineCode` on the top-level error object in
addition to where it was before on the specific `CardError` type as a
minor quality of life improvement.

It seemed like a very simple change, so we didn't put in a test, but
what we forgot about at the time was that we were already doing some
weird stuff in order to extract `DeclineCode` for `CardError`.
Specifically, we'd added this special wrapper type that extended
`Error`:

``` go
// rawErrorInternal embeds Error to deserialize all the standard error fields,
// but also adds other fields that may or may not be present depending on error
// type to help with deserialization. (e.g. DeclineCode).
type rawErrorInternal struct {
	*Error
	DeclineCode *DeclineCode `json:"decline_code,omitempty"`
}
```

It added `DeclineCode` in addition to the `DeclineCode` on the embedded
`Error`. The trouble was that Go will only deserialize any given field
_once_, so `rawErrorInternal.DeclineCode` was being populated, and
`rawErrorInternal.Error.DeclineCode` was not.

I fix the problem by essentially eliminating `rawErrorInternal` and
flattening our error type structure somewhat. This results in slightly
nicer code, and also fixes deserialization of `Error.DeclineCode`.
We continue to also support `CardError.DeclineCode` for backwards
compatibility.

Fixes #1115.