Test only on Go 1.13+
Unfortunately, Testify introduced use of `errors.As` and `errors.Is`
[1], which are only available in Go 1.13+, and therefore broke use on
all older versions of Go. Speaking in a purely technical sense only,
they're in the right because only the last two Go minors are supported
at any given time (currently, 1.13 and 1.14).

We still want to support older versions of Go for now though, so here we
modify the build to only test on 1.13+. We still build the older
versions of Go which gives us decent assurance they still work. It's not
perfect, but is a decent compromise given the circumstances in that at
least they're still in our build matrix.

[1] https://blog.golang.org/go1.13-errors#TOC_3.2.