Make original list object accessible on iterators
This patch makes some tweaks to `Iter` so that it doesn't just retain
list metadata, but a reference to the entirety of the list object.

This is mainly useful so that it'd be possible to access the last API
response, which could be used for things like this:

``` go
requestID := it.CustomerList().LastResponse.RequestID
```

Currently, it's not possible to access a list operation's last response
because we only set `LastResponse` on the top level API resource (which
is the list in this case), and `Iter` provides no access to that.

I also introduce a new `ListContainer` interface that all lists inherit
through their embedded `ListMeta` just to convey more typing information
so that we don't have to use an untyped `interface{}`. I think this is
fine though because it's very nicely symmetric with what we already have
for `ListParamsContainer` which has a nearly identical implementation.

The downside is that this changes the definition of `Query`:

``` go
// Old definition
type Query func(*Params, *form.Values) ([]interface{}, ListMeta, error)

// New definition
type Query func(*Params, *form.Values) ([]interface{}, ListContainer, error)
```

This is technically a breaking change because `Query` is exported from
the `stripe` package, but it's one of those changes that's very unlikely
to affect that vast majority of users because it's not a type that you'd
ever be using most of the time.

We also have to make a slight change to every existing `List` function
to adopt the new `Query` signature. I've shown what this would look like
in the changes in `customer/client.go`. Unfortunately, I can't find many
solutions to getting access to `LastResponse` that _don't_ involve
changing every `List` implementation to some degree. Similarly, we also
add a new function to every resource's `Iter` like `CustomerList` to get
back a typed version of the list object, similar to how each already has
a function to get a typed version of its API resource like `Customer()`.

Fixes #1141.