Don't retry requests on context cancellation + a few other errors
Currently, the retry logic is written to mostly just be smart about
status codes, like if we see a 409 or 429. We fairly naively retry any
non-Stripe error that comes back from `net/http`.

As pointed out in #1184, we should at the very least handle a context
that's been cancelled or hit its deadline (retries should stop if this
is the case), so I've done that here.

I've also tweaked some of the error logic so that we don't retry on some
types of non-Stripe errors which we know will never succeed. I don't
expect this set to be completely exhaustive, but it gives us a
framework to build on in case we want to add more errors to it in the
future. The current set is inspired by go-retryablehttp.

Fixes #1184.

[1] https://github.com/hashicorp/go-retryablehttp