Put a `Deprecated` notice on `TotalCount`
We still have `TotalCount` in a list's meta information. So as to not
make a breaking change, I'm going to leave it in, but add a `Deprecated`
notice on it.

I also reword the comment because the note to "see tests" is no longer
accurate (we don't have a `total_count` sample in the suite anymore).

Replaces #1213.