Add MsgCode to Error
### Background Summary

The current `Error` struct does not include a field to store the `message_code`. This requires writing additional code to extract the `message_code` from `RawJSON` when handling specific errors, such as checking if an invoice is already paid.

### Detailed Description

When executing `POST /v1/invoice/:id/pay` on an already paid invoice, the following error is returned:

```json
{
  "error": {
    "message": "foo",
    "message_code": "invoice_already_paid",
    "request_log_url": "https://dashboard.stripe.com/test/logs/foo",
    "type": "invalid_request_error"
  }
}
```

Currently, since the `Error` struct does not have a field that includes `message_code`, it is necessary to write code to extract `message_code` from `RawJSON`.

The example of Before and After of this modification is as follows:

```go
// Before
var stripeErr *stripe.Error
if errors.As(err, &stripeErr) {
    switch stripeErr.Type {
    case stripe.ErrorTypeInvalidRequest:
        if stripeErr.LastResponse != nil {
            var responseBody map[string]interface{}
            jsonErr := json.Unmarshal(stripeErr.LastResponse.RawJSON, &responseBody)
            if jsonErr != nil {
                return jsonErr
            }
            if errorObj, ok := responseBody["error"].(map[string]interface{}); ok {
                messageCode, _ := errorObj["message_code"].(string)
                if messageCode == "invoice_already_paid" {
                    // Process for already paid invoice
                }
            }
        }
    }
}

// After
var stripeErr *stripe.Error
if errors.As(err, &stripeErr) {
    switch stripeErr.Type {
    case stripe.ErrorTypeInvalidRequest:
        if stripeErr.MsgCode == "invoice_already_paid" {
            // Process for already paid invoice
        }
    }
}
```